完成对于课程权限的开发 fix权限admin失效问题

+++ src/main/java/com/ssgroup/zelu/pojo/department/School.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class School {
 
     @TableId(type = IdType.AUTO)
-    private Integer schoolId;
+    private Long schoolId;
 
     @NotEmpty
     private String schoolName;
